Update Middleware & Custom Login

+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class UserController extends Controller
 {
@@ -96,7 +97,11 @@ class UserController extends Controller
     public function destroy($id)
     {
         $users = User::find($id);
+        if($users->id == Auth::user()->id)
+        {
+            return back()->with('error','Tidak Bisa Menghapus Diri Sendiri');
+        }
         $users->delete();
-        return redirect()->route('users.index');
+        return redirect('users.index')->with(['success','Data Berhasil Dihapus']);
     }
 }
